wip: Motor 1 hat nun verbesserte Rampe, derzeit nur eine Richtung, TODOs
Für Motor 2 implementieren
Drehrichtungswechsel implementieren

+++ Motors.h
@@ -9,9 +9,10 @@
 #define MOTORS_H_
 
 #include <Arduino.h>
+#include <stdint.h>
 
-void setMotorSpeedM1(int16_t tspeed);
-void setMotorSpeedM2(int16_t tspeed);
+void setMotorSpeedM1(int32_t tspeed);
+void setMotorSpeedM2(int32_t tspeed);
 
 
 #endif /* MOTORS_H_ */
